# internal/api/swagger/swagger.yaml
basePath: /v1
definitions:
...
  internal_api_handler.contactRequest:
    properties:
      label:
        type: string
      position:
        type: integer
      type:
        type: string
      value:
        type: string
    type: object
...
  internal_api_handler.envelope:
    properties:
      data: {}
      error:
        type: string
    type: object
...
  internal_api_handler.locationRequest:
    properties:
      address:
        type: string
      city:
        type: string
      country:
        type: string
      position:
        type: integer
      postal_code:
        type: string
      type:
        type: string
    type: object
  internal_api_handler.personRequest:
    properties:
      contacts:
        items:
          $ref: '#/definitions/internal_api_handler.contactRequest'
        type: array
      date_of_birth:
        description: '"YYYY-MM-DD" or ""'
        type: string
      gender:
        description: '"" | "male" | "female" | "rather_not_say"'
        type: string
      locations:
        items:
          $ref: '#/definitions/internal_api_handler.locationRequest'
        type: array
      name:
        type: string
      nickname:
        type: string
      other_notes:
        type: string
      relationship_type:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    name: kith
    url: https://github.com/nhymxu/kith-pms
  description: Self-hosted personal relationship manager API.
  license:
    name: MIT
  title: kith API
  version: "1.0"
paths:
...
  /people:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Comma-separated label IDs
        in: query
        name: labels
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: List people
      tags:
      - people
    post:
      consumes:
      - application/json
      parameters:
      - description: Person data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.personRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_api_handler.envelope'
            - properties:
                data:
                  properties:
                    id:
                      type: integer
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Create person
      tags:
      - people
  /people/{id}:
    delete:
      parameters:
      - description: Person ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Delete person
      tags:
      - people
    get:
      parameters:
      - description: Person ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Get person
      tags:
      - people
    put:
      consumes:
      - application/json
      parameters:
      - description: Person ID
        in: path
        name: id
        required: true
        type: integer
      - description: Person data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.personRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api_handler.envelope'
            - properties:
                data:
                  properties:
                    id:
                      type: integer
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Update person
      tags:
      - people
...
  /people/{id}/labels:
    post:
      consumes:
      - application/json
      parameters:
      - description: Person ID
        in: path
        name: id
        required: true
        type: integer
      - description: Label ID
        in: body
        name: body
        required: true
        schema:
          properties:
            label_id:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api_handler.envelope'
            - properties:
                data:
                  properties:
                    attached:
                      type: boolean
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Attach label to person
      tags:
      - people
  /people/{id}/labels/{labelID}:
    delete:
      parameters:
      - description: Person ID
        in: path
        name: id
        required: true
        type: integer
      - description: Label ID
        in: path
        name: labelID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handler.envelope'
      security:
      - CookieAuth: []
      - CSRFHeader: []
      summary: Detach label from person
      tags:
      - people
...
securityDefinitions:
  CSRFHeader:
    description: Must be set to "kith-spa" on all mutating requests (POST/PUT/PATCH/DELETE)
    in: header
    name: X-Requested-With
    type: apiKey
  CookieAuth:
    description: HttpOnly session cookie set by POST /v1/auth/login
    in: cookie
    name: kith_session
    type: apiKey
swagger: "2.0"